+++ src/components/list/list.vue
@@ -1,21 +1,36 @@
 <template>
   <div class="spr-font-main">
-    <div
-      v-if="props.searchableMenu"
-      :class="[
-        'spr-sticky spr-z-20',
-        'spr-grid spr-gap-3 spr-bg-white-50 spr-p-size-spacing-3xs',
-        'spr-border-color-weak spr-border spr-border-x-0 spr-border-b spr-border-t-0 spr-border-solid',
-      ]"
-      :style="{
-        top:
-          typeof props.stickySearchOffset === 'number'
-            ? props.stickySearchOffset + 'px'
-            : String(props.stickySearchOffset),
-      }"
-    >
-      <spr-input-search v-model="searchText" :placeholder="props.searchableMenuPlaceholder" autocomplete="off" />
-    </div>
+    <template v-if="props.searchableMenu || props.displayListItemSelected">
+      <div
+        :class="[
+          'spr-sticky spr-z-20',
+          'spr-grid spr-gap-3 spr-bg-white-50 spr-px-size-spacing-3xs spr-py-size-spacing-2xs',
+          'spr-border-color-weak spr-border spr-border-x-0 spr-border-b spr-border-t-0 spr-border-solid',
+        ]"
+        :style="{
+          top:
+            typeof props.stickySearchOffset === 'number'
+              ? props.stickySearchOffset + 'px'
+              : String(props.stickySearchOffset),
+        }"
+      >
+        <spr-input-search
+          v-if="props.searchableMenu"
+          v-model="searchText"
+          :placeholder="props.searchableMenuPlaceholder"
+          autocomplete="off"
+        />
+        <span
+          v-if="props.supportingDisplayText || props.displayListItemSelected"
+          class="spr-label-sm-medium spr-text-color-base spr-block"
+        >
+          <template v-if="props.supportingDisplayText">
+            {{ props.supportingDisplayText }}
+          </template>
+          <template v-else> {{ selectedItems.length }} Selected</template>
+        </span>
+      </div>
+    </template>
 
     <div class="spr-p-size-spacing-3xs">
       <template v-if="props.groupItemsBy">
@@ -38,7 +53,11 @@
               >
                 <template v-if="props.lozenge">
                   <div class="spr-flex spr-items-center spr-gap-1">
-                    <spr-checkbox v-if="props.multiSelect" :disabled="item.disabled" :checked="isItemSelected(item)" />
+                    <spr-checkbox
+                      v-if="props.multiSelect"
+                      :disabled="item.disabled || (props.disabledUnselectedItems && !isItemSelected(item))"
+                      :checked="isItemSelected(item)"
+                    />
                     <spr-lozenge
                       v-if="props.lozenge"
                       :label="item.text || (item.lozengeProps?.label as string)"
@@ -52,23 +71,45 @@
                 </template>
                 <template v-else>
                   <div class="spr-flex spr-items-center spr-gap-1">
-                    <spr-checkbox v-if="props.multiSelect" :disabled="item.disabled" :checked="isItemSelected(item)" />
+                    <spr-checkbox
+                      v-if="props.multiSelect"
+                      :disabled="item.disabled || (props.disabledUnselectedItems && !isItemSelected(item))"
+                      :checked="isItemSelected(item)"
+                    />
                     <div :class="[item.textColor, 'spr-flex spr-flex-row spr-items-center spr-gap-size-spacing-3xs']">
-                      <span v-if="props.itemIcon || item.icon" :class="[item.iconColor, 'spr-mt-[2px]']">
+                      <span
+                        v-if="props.itemIcon || item.icon"
+                        :class="[
+                          item.iconColor,
+                          'spr-mt-[2px]',
+                          {
+                            'spr-text-color-disabled':
+                              item.disabled || (props.disabledUnselectedItems && !isItemSelected(item)),
+                          },
+                        ]"
+                      >
                         <icon :icon="(props.itemIcon || item.icon) as string" width="20px" height="20px" />
                       </span>
                       <div
                         :class="[
                           'spr-flex spr-flex-auto spr-flex-col spr-justify-start',
-                          { 'spr-text-color-disabled': item.disabled },
+                          {
+                            'spr-text-color-disabled':
+                              item.disabled || (props.disabledUnselectedItems && !isItemSelected(item)),
+                          },
                         ]"
                       >
-                        <span class="spr-text-left spr-text-xs">{{ item.text }}</span>
+                        <span class="spr-break-words spr-text-left spr-text-xs">
+                          {{ item.text }}
+                        </span>
                         <span
                           v-if="item.subtext"
                           :class="[
-                            'spr-body-xs-regular spr-text-color-base spr-text-left',
-                            { 'spr-text-color-disabled': item.disabled },
+                            'spr-body-xs-regular spr-text-color-base spr-break-words spr-text-left',
+                            {
+                              'spr-text-color-disabled':
+                                item.disabled || (props.disabledUnselectedItems && !isItemSelected(item)),
+                            },
                           ]"
                         >
                           {{ item.subtext }}
@@ -128,7 +169,11 @@
           >
             <template v-if="props.lozenge">
               <div class="spr-flex spr-items-center spr-gap-1">
-                <spr-checkbox v-if="props.multiSelect" :disabled="item.disabled" :checked="isItemSelected(item)" />
+                <spr-checkbox
+                  v-if="props.multiSelect"
+                  :disabled="item.disabled || (props.disabledUnselectedItems && !isItemSelected(item))"
+                  :checked="isItemSelected(item)"
+                />
                 <spr-lozenge
                   v-if="props.lozenge"
                   :label="item.text || (item.lozengeProps?.label as string)"
@@ -142,23 +187,45 @@
             </template>
             <template v-else>
               <div class="spr-flex spr-items-center spr-gap-1">
-                <spr-checkbox v-if="props.multiSelect" :disabled="item.disabled" :checked="isItemSelected(item)" />
+                <spr-checkbox
+                  v-if="props.multiSelect"
+                  :disabled="item.disabled || (props.disabledUnselectedItems && !isItemSelected(item))"
+                  :checked="isItemSelected(item)"
+                />
                 <div :class="[item.textColor, 'spr-flex spr-flex-row spr-items-center spr-gap-size-spacing-3xs']">
-                  <span v-if="props.itemIcon || item.icon" :class="[item.iconColor, 'spr-mt-[2px]']">
+                  <span
+                    v-if="props.itemIcon || item.icon"
+                    :class="[
+                      item.iconColor,
+                      'spr-mt-[2px]',
+                      {
+                        'spr-text-color-disabled':
+                          item.disabled || (props.disabledUnselectedItems && !isItemSelected(item)),
+                      },
+                    ]"
+                  >
                     <icon :icon="(props.itemIcon || item.icon) as string" width="20px" height="20px" />
                   </span>
                   <div
                     :class="[
                       'spr-flex spr-flex-auto spr-flex-col spr-justify-start',
-                      { 'spr-text-color-disabled': item.disabled },
+                      {
+                        'spr-text-color-disabled':
+                          item.disabled || (props.disabledUnselectedItems && !isItemSelected(item)),
+                      },
                     ]"
                   >
-                    <span class="spr-text-left spr-text-xs">{{ item.text }}</span>
+                    <span class="spr-break-words spr-text-left spr-text-xs">
+                      {{ item.text }}
+                    </span>
                     <span
                       v-if="item.subtext"
                       :class="[
-                        'spr-body-xs-regular spr-text-color-base spr-text-left',
-                        { 'spr-text-color-disabled': item.disabled },
+                        'spr-body-xs-regular spr-text-color-base spr-break-words spr-text-left',
+                        {
+                          'spr-text-color-disabled':
+                            item.disabled || (props.disabledUnselectedItems && !isItemSelected(item)),
+                        },
                       ]"
                     >
                       {{ item.subtext }}
@@ -225,6 +292,13 @@ import { useList } from './use-list';
 const props = defineProps(listPropTypes);
 const emit = defineEmits(listEmitTypes);
 
-const { searchText, localizedMenuList, groupedMenuList, isItemSelected, getListItemClasses, handleSelectedItem } =
-  useList(props, emit);
+const {
+  selectedItems,
+  searchText,
+  localizedMenuList,
+  groupedMenuList,
+  isItemSelected,
+  getListItemClasses,
+  handleSelectedItem,
+} = useList(props, emit);
 </script>