+++ src/components/select/select-multiple/select-multiple.ts
@@ -123,7 +123,19 @@ export const multiSelectPropTypes = {
     type: String,
     default: '',
   },
+  supportingDisplayText: {
+    type: String,
+    default: '',
+  },
+  persistentDisplayText: {
+    type: Boolean,
+    default: false,
+  },
   displaySelectedCountOnly: {
+    type: Boolean,
+    default: false,
+  },
+  displayListItemSelected: {
     type: Boolean,
     default: false,
   },
@@ -179,14 +191,22 @@ export const multiSelectPropTypes = {
     type: String,
     default: '',
   },
+  disabledUnselectedItems: {
+    type: Boolean,
+    default: false,
+  },
 };
 
 export const multiSelectEmitTypes = {
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
   'update:modelValue': (_value: unknown) => true,
   'update:searchValue': (value: string) => typeof value === 'string',
+  'search-string': (value: string) => typeof value === 'string',
   'infinite-scroll-trigger': Boolean,
   'popper-state': Boolean,
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  'get-selected-options': (_value: unknown) => true,
+  'get-single-selected-item': (item: MenuListType) => item,
 };
 
 export type MultiSelectPropTypes = ExtractPropTypes<typeof multiSelectPropTypes>;