+++ docs/guide/changelog.md
@@ -1,9 +1,19 @@
 # Changelog
 
+## 2.19.4 (2025-10-23)
+
+- Fix:
+  - Fix the issue encountered in the disabled local search list rendering reactiveness.
+  - Added support for multi-select persistent display text
+  - Added support for multi-select display selected list item.
+  - Added support for multi-select get-selected-items emit.
+  - Added support for multi-select get-single-selected-item emit.
+    ([#35548](https://dev.azure.com/sproutphil/Sprout%20Design%20System/_git/Sprout%20Design%20System%20Next/pullrequest/35548?_a=files) by @jarias)
+
 ## 2.19.3 (2025-10-23)
 
 - Feat: Added writable input support to Ladderized Select components
-    ([#e7a5b4f9](https://dev.azure.com/sproutphil/Sprout%20Design%20System/_git/Sprout%20Design%20System%20Next/commit/e7a5b4f94261d868fbe61baf2e62b83adcb8dd6d?refName=refs/heads/2025/feat/Paulo/UpdatedLadderizedListAndSelectToEnableWritableInput) by @jbanares)
+  ([#e7a5b4f9](https://dev.azure.com/sproutphil/Sprout%20Design%20System/_git/Sprout%20Design%20System%20Next/commit/e7a5b4f94261d868fbe61baf2e62b83adcb8dd6d?refName=refs/heads/2025/feat/Paulo/UpdatedLadderizedListAndSelectToEnableWritableInput) by @jbanares)
 
 ## 2.19.2 (2025-10-22)
 
@@ -13,8 +23,7 @@
 ## 2.19.1 (2025-10-22)
 
 - Chore: Enhance Chips component documentation with slot customization examples
-    ([#958ee225](https://dev.azure.com/sproutphil/Sprout%20Design%20System/_git/Sprout%20Design%20System%20Next/commit/958ee225604fe7abf7f6a2be98e9fc0c0ff7eed3?refName=refs%2Fheads%2F2025%2FFIX%2FBEN%2FChips-slot) by @bsolis)
-
+  ([#958ee225](https://dev.azure.com/sproutphil/Sprout%20Design%20System/_git/Sprout%20Design%20System%20Next/commit/958ee225604fe7abf7f6a2be98e9fc0c0ff7eed3?refName=refs%2Fheads%2F2025%2FFIX%2FBEN%2FChips-slot) by @bsolis)
 
 ## 2.19.0 (2025-10-22)